refactor: change to ts project, expose definitions as object instead of raw json
