--- index.js
@@ -1,2 +1,0 @@
-import definitions from "./definitions.json";
-export default definitions;
