+++ package.json
@@ -1,9 +1,10 @@
 {
     "name": "@interlay/interbtc-types",
-    "version": "1.13.0",
+    "version": "2.0.0",
     "description": "Substrate types used in InterBTC parachain",
+    "type": "module",
     "main": "build/index.js",
-    "typings": "index.d.ts",
+    "typings": "build/index.d.ts",
     "repository": "https://github.com/interlay/interbtc-types",
     "author": "Interlay Ltd",
     "license": "Apache-2.0",
@@ -13,11 +14,10 @@
         "InterBTC"
     ],
     "scripts": {
-        "build": "webpack",
+        "build": "rm -r build && tsc -p tsconfig.json",
         "prepublish": "yarn build"
     },
     "devDependencies": {
-        "webpack": "^5.2.0",
-        "webpack-cli": "^4.2.0"
+        "typescript": "5.2.2"
     }
 }
