+++ src/index.ts
@@ -1,0 +1,2 @@
+export { definitions } from "./definitions";
+export { RpcFunctionDefinition } from "./types";
