+++ tsconfig.json
@@ -1,0 +1,22 @@
+{
+    "ts-node": {
+        "experimentalSpecifierResolution": "node"
+    },
+    "compilerOptions": {
+        "skipLibCheck": true,
+        "target": "ES6",
+        "module": "ES6",
+        "outDir": "build",
+        "declaration": true,
+        "strict": true,
+        "noImplicitAny": true,
+        "esModuleInterop": true,
+        "baseUrl": "./",
+        "resolveJsonModule": true,
+        // Generate files for debugging
+        "sourceMap": true,
+        "moduleResolution": "node",
+    },
+    "include": ["src/**/*"],
+    "exclude": ["./node_modules/*"]
+}
