--- webpack.config.js
@@ -1,11 +1,0 @@
-const path = require("path");
-
-module.exports = {
-  entry: "./index.js",
-  output: {
-    filename: "index.js",
-    path: path.resolve(__dirname, "build"),
-    libraryTarget: "umd",
-    globalObject: "this",
-  },
-};
